Move interop to an appropriate module. Renaming.
This way the consumer of the interop code is less involved in what
occurs to make interop happen, and the implementation of the Deref
traits allows the use to be relatively transparent. Move away from the
"Wrapper" naming convention for this reason, as it implies that there's
more to drill down into, which is largely incorrect.